+++ packages-next/fields/src/types/text/views/index.tsx
@@ -86,7 +86,7 @@ export const controller = (
               props.onChange(event.target.value);
             }}
             value={props.value}
-            autoFocus
+            autoFocus={props.autoFocus}
           />
         );
       },