+++ tests/api-tests/relationships/nested-mutations/create-and-connect-singular.test.ts
@@ -1,58 +1,52 @@
-import { testConfig } from '@keystone-next/test-utils-legacy';
 import { text, relationship } from '@keystone-next/fields';
 import { createSchema, list } from '@keystone-next/keystone/schema';
-import { multiAdapterRunners, setupFromConfig } from '@keystone-next/test-utils-legacy';
-import { DatabaseProvider } from '@keystone-next/types';
+import { setupTestRunner } from '@keystone-next/testing';
+import { apiTestConfig } from '../../utils';
 
-function setupKeystone(provider: DatabaseProvider) {
-  return setupFromConfig({
-    provider,
-    config: testConfig({
-      lists: createSchema({
-        Group: list({
-          fields: {
-            name: text(),
-          },
-        }),
-        Event: list({
-          fields: {
-            title: text(),
-            group: relationship({ ref: 'Group' }),
-          },
-        }),
+const runner = setupTestRunner({
+  config: apiTestConfig({
+    lists: createSchema({
+      Group: list({
+        fields: {
+          name: text(),
+        },
+      }),
+      Event: list({
+        fields: {
+          title: text(),
+          group: relationship({ ref: 'Group' }),
+        },
       }),
     }),
-  });
-}
+  }),
+});
 
-multiAdapterRunners().map(({ runner, provider }) =>
-  describe(`Provider: ${provider}`, () => {
-    describe('errors on incomplete data', () => {
-      test(
-        'when neither id or create data passed',
-        runner(setupKeystone, async ({ context }) => {
-          // Create an item that does the linking
-          const { errors } = await context.graphql.raw({
-            query: `
+describe('errors on incomplete data', () => {
+  test(
+    'when neither id or create data passed',
+    runner(async ({ context }) => {
+      // Create an item that does the linking
+      const { errors } = await context.graphql.raw({
+        query: `
               mutation {
                 createEvent(data: { group: {} }) {
                   id
                 }
               }`,
-          });
+      });
 
-          expect(errors).toMatchObject([
-            { message: 'Nested mutation operation invalid for Event.group<Group>' },
-          ]);
-        })
-      );
+      expect(errors).toMatchObject([
+        { message: 'Nested mutation operation invalid for Event.group<Group>' },
+      ]);
+    })
+  );
 
-      test(
-        'when both id and create data passed',
-        runner(setupKeystone, async ({ context }) => {
-          // Create an item that does the linking
-          const { data, errors } = await context.graphql.raw({
-            query: `
+  test(
+    'when both id and create data passed',
+    runner(async ({ context }) => {
+      // Create an item that does the linking
+      const { data, errors } = await context.graphql.raw({
+        query: `
               mutation {
                 createEvent(data: { group: {
                   connect: { id: "abc123"},
@@ -61,14 +55,12 @@ multiAdapterRunners().map(({ runner, provider }) =>
                   id
                 }
               }`,
-          });
+      });
 
-          expect(data?.createEvent).toBe(null);
-          expect(errors).toMatchObject([
-            { message: 'Nested mutation operation invalid for Event.group<Group>' },
-          ]);
-        })
-      );
-    });
-  })
-);
+      expect(data?.createEvent).toBe(null);
+      expect(errors).toMatchObject([
+        { message: 'Nested mutation operation invalid for Event.group<Group>' },
+      ]);
+    })
+  );
+});